refactor: move to models namespace
